auto-reload for each camera permission change
