+++ src/views/SearchView.vue
@@ -20,7 +20,6 @@
   SPDX-License-Identifier: Apache-2.0
 -->
 
-
 <template>
   <div class="home-page-container">
     <div class="left-container" :class="{ hidden: isHidden }">
@@ -45,26 +44,17 @@
         <div class="logotype-container">
           <img :src="LogotypeDPP" alt="DPP logo" />
         </div>
-
-        <v-icon
-          @click="showWelcome"
-          size="large"
-          :class="{ hidden: !isHidden }"
-          class="arrow-icon"
-          icon="mdi-arrow-right"
-        ></v-icon>
-
         <v-container class="search-page">
-          <div v-if="error" class="qr-container">
+          <div v-if="qrError" class="qr-container">
             <div class="text-container">
               <p class="text">Your camera is off.</p>
               <p class="text">Turn it on or type the ID.</p>
-              <p class="error">{{ error }}</p>
+              <p class="error">{{ qrError }}</p>
             </div>
             <SearchInput class="search-input" />
           </div>
           <v-row data-cy="qr-container">
-            <div v-if="!error">
+            <div v-if="!qrError">
               <v-col class="qr-container" cols="12" v-if="QRtoggle">
                 <div class="stream-container">
                   <v-icon
@@ -75,12 +65,7 @@
                     md
                     icon="mdi-close-thick"
                   ></v-icon>
-                  <qrcode-stream
-                    :torch="torch"
-                    class="qrcode-stream"
-                    @init="onInit"
-                    @decode="onDecode"
-                  ></qrcode-stream>
+                  <QrcodeStream v-if="QRtoggle" :key="reloadReader" />
                 </div>
               </v-col>
               <v-col cols="12" v-else class="qr-container">
@@ -107,15 +92,13 @@
   </div>
 </template>
 
-
-
 <script>
-import { QrcodeStream } from "vue3-qrcode-reader";
-import CatenaLogo from "../media/logo.png";
-import QRFrame from "../media/qrFrame.svg";
+import QrcodeStream from "../components/general/QrcodeStrem.vue";
 import BatteryScanning from "../media/battery-img.jpeg";
 import LogotypeDPP from "../media/logotypeDPP.svg";
 import SearchInput from "../components/general/SearchInput.vue";
+import { mapState } from "vuex";
+import store from "@/store/index";
 
 export default {
   name: "QRScannerView",
@@ -127,69 +110,48 @@ export default {
     return {
       isHidden: false,
       QRtoggle: false,
-      error: "",
-      decodedString: "",
+      reloadReader: 0,
     };
+  },
+  computed: {
+    ...mapState(["qrError"]),
   },
   setup() {
     SearchInput;
     return {
       BatteryScanning,
       LogotypeDPP,
-      CatenaLogo,
-      QRFrame,
     };
   },
+  created() {
+    this.checkCameraPermission();
+  },
   methods: {
+    async checkCameraPermission() {
+      try {
+        const permissionStatus = await navigator.permissions.query({
+          name: "camera",
+        });
+        permissionStatus.onchange = () => {
+          this.reloadReader += 1;
+          store.commit("setQrError", "");
+        };
+      } catch (error) {
+        console.error("Error checking camera permission:", error);
+      }
+    },
     hideWelcome() {
       this.isHidden = true;
       this.QRtoggle = true;
     },
-    showWelcome() {
-      this.isHidden = false;
-      this.QRtoggle = false;
-    },
     closeQRScanner() {
       this.QRtoggle = false;
-    },
-    openQRScanner() {
-      this.QRtoggle = true;
-    },
-    async onInit(promise) {
-      try {
-        await promise;
-      } catch (error) {
-        if (error.name === "NotAllowedError") {
-          this.error = "ERROR: you need to grant camera access permission";
-        } else if (error.name === "NotFoundError") {
-          this.error = "ERROR: no camera on this device";
-        } else if (error.name === "NotSupportedError") {
-          this.error = "ERROR: secure context required (HTTPS, localhost)";
-        } else if (error.name === "NotReadableError") {
-          this.error = "ERROR: is the camera already in use?";
-        } else if (error.name === "OverconstrainedError") {
-          this.error = "ERROR: installed cameras are not suitable";
-        } else if (error.name === "StreamApiNotSupportedError") {
-          this.error = "ERROR: Stream API is not supported in this browser";
-        } else if (error.name === "InsecureContextError") {
-          this.error =
-            "ERROR: Camera access is only permitted in secure context. Use HTTPS or localhost rather than HTTP.";
-        } else {
-          this.error = `ERROR: Camera error (${error.name})`;
-        }
-      }
     },
     openExternalLink() {
       window.open(
         "https://portal.int.demo.catena-x.net/documentation/?path=docs",
         "_blank"
       );
-    },
-    onDecode(decodedString) {
-      this.decodedString = decodedString;
-      this.$router.push({
-        path: `/${decodedString}`,
-      });
     },
   },
 };
